documentando no readme e melhorias no script
